nuplayer: Allow devices to force a flush when surface changes
This fixes some playback problems on the Zenfone2 (running on 5.0
based blobs).

The problems are easily reproduced by viewing a video in Chrome and
then switching to full-screen mode.  The change in surface causes
the log to get flooded with messages like:

dc_device_post: Buffer ID=14460 not imported by D

until Android crashes.  With this change, it ends up correctly
mapping the surfaces after the flush fixing the problem.

The code that is being
ifdef'ed out was added in M in commit
nuplayer: use codec->setSurface when possible to avoid seeking

and it appears that our blobs simply cannot handle this optimization
and therefore we need a means by which to disable it.

Change-Id: Ie644c17abeed58e7d179e1d9eb34441fea67b912