libstagefright: Make it possible to skip OMX buffer reallocation
Some devices don't like the call to setParameter() at this point, so
skip this call if enough buffers are already allocated. This check
was present in KitKat but got removed when code to allocate extra-
buffers was introduced.
This is activated only for omap4 for now, but also has an opt-in via BOARD_CANT_REALLOCATE_OMX_BUFFERS

Change-Id: I7a2f9b53905426d9bf75c44bc211103def390b4a